monitoring-java/requirements.txt
dataiku-api-client